requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-dialogflow
flask-cors
jsonpickle